Rename from filter to strategy
