+++ src/ComponentFinder.php
@@ -7,10 +7,22 @@ use SplFileInfo;
 use Illuminate\Support\Collection;
 use Symfony\Component\Finder\Finder;
 use Wnx\LaravelStats\Classifiers\Classifier;
-use Wnx\LaravelStats\Filters\RejectVendorClasses;
+use Wnx\LaravelStats\RejectionStrategies\RejectVendorClasses;
 
 class ComponentFinder
 {
+    /**
+     * @var RejectionStrategy
+     */
+    protected $rejectionStrategy;
+
+    public function __construct()
+    {
+        $rejectionStrategy = config('stats.rejection_strategy', RejectVendorClasses::class);
+
+        $this->rejectionStrategy = resolve($rejectionStrategy);
+    }
+
     /**
      * Sort classes into Laravel Component.
      *
@@ -23,9 +35,7 @@ class ComponentFinder
                 return new ReflectionClass($class);
             })
             ->reject(function ($class) {
-                $filter = config('stats.filter', RejectVendorClasses::class);
-
-                return (resolve($filter))->shouldBeRejected($class);
+                return $this->rejectionStrategy->shouldClassBeRejected($class);
             })
             ->groupBy(function ($class) {
                 return (new Classifier)->classify($class);
