--- src/Contracts/Filter.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace Wnx\LaravelStats\Contracts;
-
-use Wnx\LaravelStats\ReflectionClass;
-
-interface Filter
-{
-    public function shouldBeRejected(ReflectionClass $class) : bool;
-}
