--- src/Filters/RejectVendorClasses.php
@@ -1,14 +1,0 @@
-<?php
-
-namespace Wnx\LaravelStats\Filters;
-
-use Wnx\LaravelStats\ReflectionClass;
-use Wnx\LaravelStats\Contracts\Filter;
-
-class RejectVendorClasses implements Filter
-{
-    public function shouldBeRejected(ReflectionClass $class) : bool
-    {
-        return $class->isInternal() || $class->isVendorProvided();
-    }
-}
