+++ src/RejectionStrategies/RejectInternalClasses.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Wnx\LaravelStats\RejectionStrategies;
+
+use Wnx\LaravelStats\ReflectionClass;
+use Wnx\LaravelStats\Contracts\RejectionStrategy;
+
+class RejectInternalClasses implements RejectionStrategy
+{
+    public function shouldClassBeRejected(ReflectionClass $class) : bool
+    {
+        return $class->isInternal();
+    }
+}
